% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{color.gradient}
\alias{color.gradient}
\title{create a gradient of colors givent a vector of value}
\usage{
color.gradient(x, colors = c(myred, "yellow", mygreen), colsteps = 100)
}
\arguments{
\item{x}{a vector of value to be matched with colors}

\item{colors}{a vector of three colors}

\item{colsteps}{the number of cilors groups}
}
\description{
create a gradient of colors givent a vector of value
}
