% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abmEpi.R
\name{generatePopulation}
\alias{generatePopulation}
\title{Population initiliasation}
\usage{
generatePopulation(N, agedistrib = NULL, behavior = NULL,
  xsize = 100, ysize = 100, recovery = NULL, speed = NULL)
}
\arguments{
\item{N}{number of agents}

\item{agedistrib}{a distribution defining the percentage of population represented in different age class}

\item{behavior}{a distribution of behaviors}

\item{xsize}{spatial limits (to keep in the model?)}

\item{ysize}{spatial limits}

\item{recovery}{a distribution recovery time}
}
\description{
generate a population
}
