% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slsirSimu.R
\name{sig}
\alias{sig}
\title{Sigmoid}
\usage{
sig(x, a = 10, b = 0.5)
}
\arguments{
\item{x}{vector of value}

\item{a}{parameter to define the steepness of the sigmoid slope}

\item{b}{parameter to define when the slope starts}
}
\description{
sigmoid function for individual learning.
}
