% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slsirSimu.R
\name{slsirSimu}
\alias{slsirSimu}
\title{ABM SIR}
\usage{
slsirSimu(
  pop,
  tstep,
  p = 1,
  i0 = 1,
  di = 2,
  recovery = 10,
  speed = 0.8,
  xsize = 100,
  ysize = 100,
  visu = FALSE,
  inf = 0.5,
  sat = 10,
  sat_r = 10000,
  inf_r = 1.1,
  log = F,
  checkcountact = F,
  ts = T,
  ap = F,
  p_i = 1,
  foldername = F,
  strategy = "all",
  sl_rad = 10,
  bt = 0,
  recover = 10
)
}
\arguments{
\item{tstep}{the duration of the simualtion}

\item{p}{the probability for one agent to transmit the disease to another one. This can be a single value, in that case the probability is the same for all agent, or a vector of probabilities, in which case the vector represents the probability of transmission for different behaviors. If the p is a vector a column of bheavior should be present in pop.}

\item{i0}{the number of initial infections}

\item{di}{the distance between two agents under which the disease can be transmitted}

\item{speed}{the speed at which agents move}

\item{visu}{TRUE or FALSE, if output should be plotted}

\item{inf}{inflexion point of the sigmoid}

\item{sat}{speed of saturation of the sigmoid}

\item{log}{TRUE or FALSE, if a simple log (number of time step) should be outputed}

\item{ts}{count and return the number of users during the run}

\item{ap}{keep and return the full population for each time step, this can slow the model a lot}

\item{p_i}{the probability for one agent to learn individually.}

\item{foldername}{if a string, the output is written in a file (instead of being show in the screen) and the string is used as the name of the folder}

\item{strategy}{which social learning will be used (ie which agents will be copied)}

\item{bt}{burn in period time during whicih there is no social learning}

\item{N}{the number of agents}

\item{reco}{time for agent to Recover}
}
\description{
Main model
}
