% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{visualize}
\alias{visualize}
\title{Output}
\usage{
visualize(allpop, timeseries, xsize, ysize, foldername = FALSE,
  width = 1200, height = 900)
}
\arguments{
\item{allpop}{a list of populations for a serie of timesteps}

\item{foldername}{if a string, the output is written in a file (instead of being show in the screen) and the string is used as the name of the folder}

\item{width}{width of output file}

\item{height}{height of output file}
}
\description{
plot population and other info
}
